Add dumb vim utilities lib
I've been using VIM off and on for years to keep a work journal so this
lib has some brain-dead things for adding entries. There are few other
things that don't really get used anymore as well.

The most recent is FormatJSON(). It pretty prints JSON messages embedded
in single lines in OpenStack logs. Kind of handy for analyzing problems.